refactor: remove duplication
